refactor(io): change buffer allocation size parameter from Long to Int

+++ src/main/kotlin/com/lyni/udsonk/common/io/DirectBufferPool.kt
@@ -10,14 +10,14 @@ object DirectBufferPool {
     private val pool = LinkedList<ByteBuffer>()
 
     @JvmStatic
-    fun allocate(size: Long): ByteBuffer {
+    fun allocate(size: Int): ByteBuffer {
         pool.forEach {
             if (it.capacity() >= size) {
                 pool.remove(it)
-                return it.clear().limit(size.toInt())
+                return it.clear().limit(size)
             }
         }
-        return ByteBuffer.allocate(size.toInt())
+        return ByteBuffer.allocate(size)
     }
 
     @JvmStatic
